refactor: rename parse_file_tree
- Rename parse_file_tree.py to file_tree_parser.py
